requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
streamlit==1.27.2
requests==2.31.0
python-dotenv==1.0.0
langchain==0.0.350
langchain-deepseek==0.0.3
pyyaml==6.0.1
difflib3==0.1.5 